# 102022400280_Data_Mahasiswa_Service/storage/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student Service API",
        "description": "API untuk data mahasiswa dan validasi kuota",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get all active students",
                "operationId": "3a8420677e3657ec19c6b44fe2aac089",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get student detail by ID",
                "operationId": "d81f6ea5b7ce2df029baf28ebc4e63ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/students/validate-quota": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Validate student quota",
                "operationId": "893750f6de6c0b00cfc870c88465f14b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "requested_sks"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "requested_sks": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation success"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Student Service API"
        }
    ]
}